Maybe my very special Input and OutputStream, that loads the local stored file-blocks into its buffer, is the fast magic of hadoop?!

+++ dxram_part/dxapp/src/main/java/de/hhu/bsinfo/app/DxramFsCore/Blockinfo.java
@@ -4,8 +4,8 @@ import java.util.Arrays;
 
 public class Blockinfo {
     public long ID; // chunkid of that BlockinfoChunk
-    public long offset;  // @todo Offset of the block in the file (do I need this? Did I need it for append?)
-    public long length; // @todo home many byte did we need from this block?
+    public int offset;  // @todo Offset of the block in the file (do I need this? Did I need it for append?)
+    public int length; // @todo how many byte did we need from this block?
     public boolean corrupt; // @todo how does this happend?
     public long storageId; // to the BlockChunk id, where the data exists (only 1 id because no replica)
     /* we do not have replica like in the blocklocation class of hadoop */
